Corrected stepping into the debugger. Reorganized.
When tracing is removed from a procedure it apparently can not be turned
back on again. Not sure why. So I set a variable that indicates if the
debugger has dived into it's self due to variable tracing. When the
variable trace callback finishes the indicator is reset.

Moved the utility procedures to the top of the file.